+++ src/aoc/days/day3/IntersectionComparator.java
@@ -9,6 +9,6 @@ public class IntersectionComparator implements Comparator<Intersection> {
     @Override
     public int compare(Intersection i1, Intersection i2) {
         Point2D origo = new Point2D(0, 0);
-        return i1.getPoint().distanceFrom(origo) - i2.getPoint().distanceFrom(origo);
+        return (int)(i1.getPoint().distanceFrom(origo) - i2.getPoint().distanceFrom(origo));
     }
 }
