+++ src/aoc/days/day3/StepComparator.java
@@ -5,6 +5,6 @@ import java.util.Comparator;
 public class StepComparator implements Comparator<Intersection> {
     @Override
     public int compare(Intersection i1, Intersection i2) {
-        return i1.getSteps() - i2.getSteps();
+        return (int)(i1.getSteps() - i2.getSteps());
     }
 }
